Remove Schedule since it is superfluous
The League#schedule could just as well be an array of Rounds and remove
a level of indirection that isn't needed. This matches Yaml better
anyways.
