+++ league.js
@@ -1,17 +1,17 @@
 const { Coach } = require('./coaches.js');
-const { Schedule } = require('./schedule.js');
 
 class League {
     constructor(data) {
         this.name = data.name;
         this.ownerId = data.ownerId;
         this.currentRound = data.currentRound;
-        this.coaches = data.coaches.map(makeCoach);
-        this.schedule = new Schedule(data.schedule, this.coaches);
+        this.coaches = data.coaches.map((d) => new Coach(d));
+        this.schedule = data.schedule.map((d) => new Round(d, this.coaches));
+        this.schedule.sort((l, r) => { l.id - r.id; });
     }
 
     getCurrentRound() {
-        return this.schedule.rounds[this.currentRound - 1];
+        return this.schedule[this.currentRound - 1];
     }
 
     get ownerMentionString() {
@@ -19,9 +19,24 @@ class League {
     }
 }
 
-// Mappable constructor
-function makeCoach(data) {
-    return new Coach(data);
+class Round {
+    constructor(data, coaches) {
+        this.id = data.round;
+        this.games = data.games.map((d) => new Game(d, coaches));
+    }
 }
 
-module.exports = { League };
+class Game {
+    constructor(data, coaches) {
+        this.home = data.home;
+        this.away = data.away;
+        this.homeCoach = coaches.find(c => c.teamName === data.home)
+        this.awayCoach = coaches.find(c => c.teamName === data.away)
+    }
+
+    get coaches() {
+        return [this.homeCoach, this.awayCoach];
+    }
+}
+
+module.exports = { League, Round, Game };
